Add util unit tests
