+++ go.mod
@@ -6,4 +6,5 @@ require (
 	github.com/fluxcd/flux2 v0.2.1
 	github.com/hashicorp/terraform-plugin-sdk/v2 v2.0.4
 	github.com/pkg/errors v0.9.1 // indirect
+	github.com/stretchr/testify v1.4.0
 )
